pyq_pipeline/requirements.txt
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
supabase==2.10.0
streamlit>=1.28.0
plotly>=5.17.0
